chore: Extract localization warning helpers
